Implemented Scene manager & scene set
The Scene set is a prefab which has 2 arrays, one for bg/mg/fg and one for assets. Right now, there are 2 in the active sample scene. One for oasis one for Tropic.
Scene manager object tracks the arduino listener, as well as the 3 sets of objects that represent the background, midground, and foreground layer. It also has a public field for the starting sceneSet. This will eventually be swapped and updated using the arduino input.